add logout to auth service
