+++ docker-compose.yml
@@ -14,7 +14,7 @@ services:
       REACT_APP_AUTH_COGNITO_AUTHORIZE_URL: 'https://bwell-dev.auth.us-east-1.amazoncognito.com/oauth2/authorize'
       REACT_APP_AUTH_COGNITO_TOKEN_URL: 'https://bwell-dev.auth.us-east-1.amazoncognito.com/oauth2/token'
       REACT_APP_AUTH_COGNITO__REDIRECT_URL: 'http://localhost:5051/authcallback'
-      REACT_APP_AUTH_COGNITO_LOGOUT_URL: 'https://bwell-dev.auth.us-east-1.amazoncognito.com/oauth2/logout'
+      REACT_APP_AUTH_COGNITO_LOGOUT_URL: 'https://bwell-dev.auth.us-east-1.amazoncognito.com/logout'
       REACT_APP_AUTH_COGNITO_WELL_KNOWN_URL: 'https://cognito-idp.us-east-1.amazonaws.com/us-east-1_yV7wvD4xD/.well-known/openid-configuration'
       REACT_APP_AUTH_COGNITO_TOKEN_TO_SEND_TO_FHIR_SERVER: 'jwt' # other option is 'id_token'
       # Each AuthProvider has its own set of environment variables
